+++ packages/headless/src/app/headless-engine.test.ts
@@ -10,6 +10,11 @@ import {AnalyticsClientSendEventHook} from 'coveo.analytics/dist/definitions/cli
 import pino from 'pino';
 import {NoopPreprocessRequestMiddleware} from '../api/platform-client';
 import {validatePayloadAndThrow} from '../utils/validate-payload';
+import {
+  NoopPostprocessFacetSearchResponseMiddleware,
+  NoopPostprocessQuerySuggestResponseMiddleware,
+  NoopPostprocessSearchResponseMiddleware,
+} from '../api/search/search-api-client-middleware';
 
 describe('headless engine', () => {
   let options: HeadlessOptions<typeof searchAppReducers>;
@@ -26,6 +31,9 @@ describe('headless engine', () => {
           logger,
           renewAccessToken: async () => '',
           preprocessRequest: NoopPreprocessRequestMiddleware,
+          postprocessSearchResponseMiddleware: NoopPostprocessSearchResponseMiddleware,
+          postprocessQuerySuggestResponseMiddleware: NoopPostprocessQuerySuggestResponseMiddleware,
+          postprocessFacetSearchResponseMiddleware: NoopPostprocessFacetSearchResponseMiddleware,
         }),
         analyticsClientMiddleware: {} as AnalyticsClientSendEventHook,
         logger: logger,