+++ packages/headless/src/features/redirection/redirection-slice.test.ts
@@ -16,6 +16,11 @@ import {SearchAppState} from '../../state/search-app-state';
 import {NoopPreprocessRequestMiddleware} from '../../api/platform-client';
 import pino from 'pino';
 import {validatePayloadAndThrow} from '../../utils/validate-payload';
+import {
+  NoopPostprocessFacetSearchResponseMiddleware,
+  NoopPostprocessQuerySuggestResponseMiddleware,
+  NoopPostprocessSearchResponseMiddleware,
+} from '../../api/search/search-api-client-middleware';
 
 describe('redirection slice', () => {
   const logger = pino({level: 'silent'});
@@ -60,6 +65,9 @@ describe('redirection slice', () => {
       renewAccessToken: async () => '',
       logger,
       preprocessRequest: NoopPreprocessRequestMiddleware,
+      postprocessSearchResponseMiddleware: NoopPostprocessSearchResponseMiddleware,
+      postprocessFacetSearchResponseMiddleware: NoopPostprocessFacetSearchResponseMiddleware,
+      postprocessQuerySuggestResponseMiddleware: NoopPostprocessQuerySuggestResponseMiddleware,
     });
     const triggers = trigger ? [trigger] : [];
     jest.spyOn(apiClient, 'plan').mockResolvedValue({